textplexus/requirements.txt
langgraph
langchain
langchain-google-genai
sqlalchemy
aiosqlite
pydantic
pandas
sympy
numpy
unstructured
chromadb
python-dotenv
duckduckgo-search
streamlit
